pyproject.toml
[tool.poetry]
name = "biomed-kg-agent"
version = "0.1.0"
description = "Biomedical Knowledge Graph with Agentic Reasoning"
authors = ["Max Sajitz <msajitz@posteo.eu>"]

packages = [
    { include = "biomed_kg_agent", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0"
neo4j = "^5.14.0"
scispacy = "^0.5.2"
spacy = "^3.7.2"
langchain = "^0.3.13"
langchain-anthropic = "^0.3.0"
langgraph = "^0.2.30"
pydantic = "^2.6.0"
streamlit = "^1.40.0"
requests = "^2.31.0"
aiohttp = "^3.11.18"
sqlmodel = "^0.0.24"
# Fast models for cancer metabolomics (CI-friendly)
en-ner-bionlp13cg-md = {url = "https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_ner_bionlp13cg_md-0.5.4.tar.gz"}
en-ner-bc5cdr-md = {url = "https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_ner_bc5cdr_md-0.5.4.tar.gz"}
en-ner-craft-md = {url = "https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_ner_craft_md-0.5.4.tar.gz"}
# Entity linking model (includes UMLS linker)
en-core-sci-sm = {url = "https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_sm-0.5.4.tar.gz"}
# en-core-sci-md = {url = "https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_md-0.5.4.tar.gz"}
# en-core-sci-lg = {url = "https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_lg-0.5.4.tar.gz"}
# en-core-sci-scibert = {url = "https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_scibert-0.5.4.tar.gz"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
black = "^24.3.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
pre-commit = "^3.6.0"
mypy = "^1.10.0"
coverage = "^7.4.1"
pytest-cov = "^5.0.0"
pydantic-settings = "^2.9.1"
pytest-mock = "^3.14.1"
types-requests = "^2.32.0.20250515"
types-pyyaml = "^6.0.12.20250516"

[tool.poetry.group.notebook.dependencies]
ipython = "^9.2.0"
jupyterlab = "^4.4.3"
ipykernel = "^6.29.5"
pandas = "^2.2.3"
matplotlib = "^3.10.3"
nltk = "^3.8.1"
python-dotenv = "^1.0.1"
plotly = "^6.5.2"

[tool.poetry.group.large-models.dependencies]
# Large models for production/research use (not CI) - REMOVED TO SAVE DISK SPACE
# en-core-sci-scibert = {url = "https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_scibert-0.5.4.tar.gz"}

[tool.poetry.scripts]
biomed-agent = "biomed_kg_agent.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
